+++ cmd/zpool/zpool_vdev.c
@@ -84,7 +84,7 @@
 #include <blkid/blkid.h>
 #include "zpool_util.h"
 #include <sys/zfs_context.h>
-#include <sys/vdev_draid_impl.h>
+#include "draid_config.h"
 
 /*
  * For any given vdev specification, we can have multiple errors.  The
